:hammer: Extract & refactor
