+++ src/containers/App/Header.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const Header = styled.header`
+  padding: 4rem 0 2rem;
+  text-align: center;
+
+  p {
+    font-size: 1.25rem;
+    font-weight: 300;
+  }
+`;
+
+export default Header;
