+++ src/containers/App/Nav.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+const Nav = styled.nav`
+  ul {
+    list-style: none;
+    text-align: center;
+  }
+
+  li {
+    margin: 0.5rem;
+    display: inline-block;
+  }
+`;
+
+export default Nav;
