--- src/containers/Counter/Counter.js
@@ -1,185 +1,0 @@
-import React, { Component, PropTypes } from 'react';
-import { connect } from 'react-redux';
-import { increment, decrement, incrementIfOdd, incrementAsync } from 'actions/counter';
-import styled, { keyframes } from 'styled-components';
-import { hexToRgba } from 'styles/utils/color';
-
-const ripple = keyframes`
-  0% {
-    transform: translate(-50%, -50%) scale(0.5);
-    opacity: 1;
-  }
-
-  100% {
-    transform: translate(-50%, -50%) scale(2.2);
-    opacity: 0;
-  }
-`;
-
-const pulse = keyframes`
-  0% {
-    transform: scale(1);
-  }
-
-  20% {
-    transform: scale(1.08);
-  }
-
-  30% {
-    transform: scale(1.1);
-  }
-
-  60% {
-    transform: scale(0.85);
-  }
-
-  100% {
-    transform: scale(1);
-  }
-`;
-
-const Actions = styled.div`
-  button {
-    margin: 0.5rem;
-    padding: 0.25rem 0.8rem;
-    border: 1px solid ${props => props.theme.colors.primary};
-    border-radius: 2px;
-    color: ${props => props.theme.colors.primary};
-    font-weight: 300;
-
-    &:hover {
-      background: ${props => hexToRgba(props.theme.colors.primary, 0.1)};
-    }
-  }
-`;
-
-const Button = styled.button`
-  position: relative;
-  display: inline-block;
-  width: 5.5rem;
-  height: 5.5rem;
-  margin-top: 3rem;
-  margin-bottom: 3.5rem;
-  font-size: 2rem;
-  border-radius: 50%;
-  background: ${props => props.theme.colors.primary};
-
-  span {
-    position: relative;
-    z-index: 1;
-  }
-
-  &::before,
-  &::after {
-    content: "";
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    width: 100%;
-    height: 100%;
-    border-radius: 50%;
-    transform: translate(-50%, -50%) scale(0.5);
-  }
-`;
-
-const Base = styled.div`
-  margin: 0 auto;
-  width: 100%;
-  max-width: 768px;
-  text-align: center;
-`;
-
-const Animating = styled.div`
-  margin: 0 auto;
-  width: 100%;
-  max-width: 768px;
-  text-align: center;
-
-  ${Button} {
-    animation: ${pulse} 0.5s ease-out forwards;
-
-    &::before {
-      background: ${props => hexToRgba(props.theme.colors.primary, 0.3)};
-      animation: ${ripple} 0.8s 0.2s forwards;
-    }
-
-    &::after {
-      background: ${props => hexToRgba(props.theme.colors.primary, 0.7)};
-      animation: ${ripple} 0.8s 0.5s forwards;
-    }
-  }
-`;
-
-class Counter extends Component {
-  static propTypes = {
-    counter: PropTypes.number.isRequired,
-    dispatch: PropTypes.func.isRequired,
-  };
-
-  constructor() {
-    super();
-    this.decrement = this.decrement.bind(this);
-    this.increment = this.increment.bind(this);
-    this.incrementIfOdd = this.incrementIfOdd.bind(this);
-    this.incrementAsync = this.incrementAsync.bind(this);
-    this.animationDone = this.animationDone.bind(this);
-    this.state = {
-      animating: false,
-    };
-  }
-
-  decrement() {
-    this.props.dispatch(decrement());
-  }
-
-  increment() {
-    this.setState({ animating: true });
-    this.props.dispatch(increment());
-  }
-
-  incrementIfOdd() {
-    this.props.dispatch(incrementIfOdd());
-  }
-
-  incrementAsync() {
-    this.props.dispatch(incrementAsync(2000));
-  }
-
-  animationDone() {
-    this.setState({ animating: false });
-  }
-
-  render() {
-    const { counter } = this.props;
-
-    let Wrapper = Base;
-    if (this.state.animating) {
-      Wrapper = Animating;
-    }
-
-    return (
-      <Wrapper>
-        <Button
-          onClick={this.increment}
-          onMouseUp={this.animationDone}
-        >
-          <span>{counter}</span>
-        </Button>
-        <Actions>
-          <button onClick={this.increment} onMouseUp={this.animationDone}> + </button>
-          <button onClick={this.decrement}> – </button>
-          <button onClick={this.incrementIfOdd}> odd </button>
-          <button onClick={this.incrementAsync}> async </button>
-        </Actions>
-      </Wrapper>
-    );
-  }
-}
-
-function mapStateToProps(state) {
-  return {
-    counter: state.counter,
-  };
-}
-
-export default connect(mapStateToProps)(Counter);
