+++ src/containers/NotFound/H2.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const H2 = styled.h2`
+  margin-top: 2.5rem;
+  font-size: 1.5rem;
+  font-weight: 400;
+  letter-spacing: 0.1rem;
+`;
+
+export default H2;
