+++ src/containers/NotFound/Wrapper.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const Wrapper = styled.div`
+  margin: 0 auto;
+  width: 100%;
+  max-width: 768px;
+  text-align: center;
+`;
+
+export default Wrapper;
