+++ src/containers/NotFound/index.js
@@ -1,3 +1,11 @@
-import component from './NotFound';
+import React from 'react';
+import Wrapper from './Wrapper';
+import H2 from './H2';
 
-export default component;
+const NotFound = () => (
+  <Wrapper>
+    <H2>Page Not Found!</H2>
+  </Wrapper>
+);
+
+export default NotFound;
